+++ src/parameterupdates.h
@@ -16,7 +16,8 @@ double update_alpha(arma::vec& alpha_acceptance,
                   const std::string& metric,
                   const double& lambda,
                   const Rcpp::Nullable<arma::vec> cardinalities = R_NilValue,
-                  const Rcpp::Nullable<arma::vec> logz_estimate = R_NilValue);
+                  const Rcpp::Nullable<arma::vec> logz_estimate = R_NilValue,
+                  double alpha_max = 1e6);
 
 void update_rho(arma::cube& rho, arma::vec& rho_acceptance, arma::mat& rho_old,
                 int& rho_index, const int& cluster_index, const int& rho_thinning,